Add pike model and ability to animate it
There are a lot of global variables created in different
cells that I will get rid of in the next commit. Preserving
those cells' code is a major reason I am making this
commit.

However, I don't like having all those globals hanging
around because one caused a bug that took me several hours
yesterday and today to get rid of. (I typed `fig` instead of
`self.fig` and the editor didn't notice because of the global
variable `fig` that was the right type.)