Start writing advance_state
This is the beginning of the main physical model
(in absence of any control signals). Not much
here because I spent a lot of time deriving the
equation that gives force on the pike from
a given rope tension. (I wrote it for acute angles
but not for oblique ones yet. I haven't incorporated
either into the code yet.)